changed some directory structure, moved BB algo to modeling
