+++ doc/examples/visual/ex_visual.rst
