+++ skhep/modeling/__init__.py
@@ -1,18 +1,14 @@
 # Licensed under a 3-clause BSD style license, see LICENSE.
 """
-Module for mathematical functions and utilities.
+Module for algorithms and methods used to model distributions.
 
 This module contains in particular:
 
-  * Vector classes.
-  * Geometry classes.
+  * Bayesian Blocks binning algorithm.
 """
 
 #-----------------------------------------------------------------------------
 # Import statements
 #-----------------------------------------------------------------------------
 from __future__ import absolute_import
-
-from .kinematics import *
-from .vectors import *
-from .geometry import *
+from .bayesian_blocks import bayesian_blocks
