+++ skhep/modeling/bayesian_blocks.py
