+++ skhep/visual/mpl_plotter.py
@@ -18,7 +18,7 @@ from matplotlib.patches import Polygon
 
 import pandas as pd
 
-from skhep.math import bayesian_blocks
+from skhep.modeling import bayesian_blocks
 from skhep.visual.fill_between_steps import fill_between_steps
 
 import warnings
