+++ tests/data/visual/answers_blocks_hist.npz
