+++ tests/data/visual/answers_multi_hist1.npz
