+++ tests/data/visual/answers_simple_hist1.npz
