+++ tests/data/visual/answers_simple_hist2.npz
