+++ tests/visual/test_hists.py
@@ -1,7 +1,7 @@
 import os
 import matplotlib as mpl
 
-if os.environ.get('DISPLAY','') == '':
+if os.environ.get('DISPLAY', '') == '':
     print('no display found. Using non-interactive Agg backend')
     mpl.use('Agg')
 
@@ -10,8 +10,7 @@ import matplotlib.pyplot as plt
 import skhep
 from skhep.visual import MplPlotter as skh_plt
 
-test_dir = os.path.dirname(skhep.__file__)+'/../tests/visual'
-
+answer_dir = os.path.dirname(skhep.__file__)+'/../tests/data/visual'
 
 
 def test_simple_hist1(cmdopt, data_gen1):
@@ -19,11 +18,11 @@ def test_simple_hist1(cmdopt, data_gen1):
     output = skh_plt.hist(data_gen1[0])
 
     if cmdopt == "generate":
-        with open(test_dir+'/answers/answers_simple_hist1.npz', 'wb') as f:
+        with open(answer_dir+'/answers_simple_hist1.npz', 'wb') as f:
             np.savez(f, bc=output[0], be=output[1])
         plt.show()
     elif cmdopt == "test":
-        answers = np.load(test_dir+'/answers/answers_simple_hist1.npz')
+        answers = np.load(answer_dir+'/answers_simple_hist1.npz')
         assert(np.all(output[0] == answers['bc']))
         assert(np.all(output[1] == answers['be']))
 
@@ -33,11 +32,11 @@ def test_simple_hist2(cmdopt, data_gen1):
     output = skh_plt.hist(data_gen1[0], weights=data_gen1[2], bins=20, normed=True, color='red')
 
     if cmdopt == "generate":
-        with open(test_dir+'/answers/answers_simple_hist2.npz', 'wb') as f:
+        with open(answer_dir+'/answers_simple_hist2.npz', 'wb') as f:
             np.savez(f, bc=output[0], be=output[1])
         plt.show()
     elif cmdopt == "test":
-        answers = np.load(test_dir+'/answers/answers_simple_hist2.npz')
+        answers = np.load(answer_dir+'/answers_simple_hist2.npz')
         assert(np.all(output[0] == answers['bc']))
         assert(np.all(output[1] == answers['be']))
 
@@ -47,11 +46,11 @@ def test_blocks_hist(cmdopt, data_gen1):
     output = skh_plt.hist(data_gen1[0], bins='blocks', scale='binwidth', color='green')
 
     if cmdopt == "generate":
-        with open(test_dir+'/answers/answers_blocks_hist.npz', 'wb') as f:
+        with open(answer_dir+'/answers_blocks_hist.npz', 'wb') as f:
             np.savez(f, bc=output[0], be=output[1])
         plt.show()
     elif cmdopt == "test":
-        answers = np.load(test_dir+'/answers/answers_blocks_hist.npz')
+        answers = np.load(answer_dir+'/answers_blocks_hist.npz')
         assert(np.all(output[0] == answers['bc']))
         assert(np.all(output[1] == answers['be']))
 
@@ -61,10 +60,10 @@ def test_multi_hist1(cmdopt, data_gen1):
     output = skh_plt.hist([data_gen1[0], data_gen1[2]], bins=20, stacked=False, normed=True)
 
     if cmdopt == "generate":
-        with open(test_dir+'/answers/answers_multi_hist1.npz', 'wb') as f:
+        with open(answer_dir+'/answers_multi_hist1.npz', 'wb') as f:
             np.savez(f, bc=output[0], be=output[1])
         plt.show()
     elif cmdopt == "test":
-        answers = np.load(test_dir+'/answers/answers_multi_hist1.npz')
+        answers = np.load(answer_dir+'/answers_multi_hist1.npz')
         assert(np.all(output[0] == answers['bc']))
         assert(np.all(output[1] == answers['be']))
