+++ tox.ini
@@ -1,15 +1,13 @@
 [tox]
 envlist = py26,py27,py36
 [testenv]
-install_command = pip uninstall wheel; pip install {packages}
-list_dependencies_command = pip freeze
 deps=
-    py26: wheel==0.29.0
+    -cconstraints.txt
     pytest
     PyPDT
     py26: numpy<1.12.0
     py{27,36}: numpy
-    py26: pandas==0.16.0
+    py26: pandas==0.16.2
     py{27,36}: pandas
     py26: matplotlib<1.5.0
     py{27,36}: matplotlib<2.1.0
